# docs/backend/nivesh-app-openapi.yaml
openapi: 3.1.0
info:
  title: Nivesh App API
  version: 2.0.0
  description: |
    REST API for the Nivesh.ai application — portfolio management, goal planning,
    AI copilot, broker integration, MFD workspace, and admin controls.

    ## Authentication

    All endpoints (except `/api/auth/*`, `/api/public/*`, and `/api/cas-parser-provider/active`)
    require a session cookie set by `POST /api/auth/google`.

    Admin endpoints additionally require the requesting user to have `is_admin = true`.

    ## Base URL

    - **Production**: `https://niveshcopilot.com`
    - **Staging**: `https://staging.niveshcopilot.com`

  contact:
    name: Nivesh
    url: https://nivesh.com
servers:
  - url: https://niveshcopilot.com
    description: Production
  - url: https://staging.niveshcopilot.com
    description: Staging
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: Session cookie set by `POST /api/auth/google`.
  parameters:
    portfolio_id_query:
      name: portfolio_id
      in: query
      schema:
        type: string
      description: Filter by portfolio ID. Empty string = all portfolios.
...
  schemas:
...
    Holding:
      type: object
      properties:
        holding_id:
          type: string
        portfolio_id:
          type: string
        name:
          type: string
        ticker:
          type: string
          nullable: true
        asset_type:
          type: string
        quantity:
          type: number
        buy_price:
          type: number
        current_price:
          type: number
        sector:
          type: string
          nullable: true
        buy_date:
          type: string
          format: date
          nullable: true
...
security:
  - SessionCookie: []
paths:
...
  /api/portfolio/holdings:
    get:
      tags: [portfolio]
      operationId: holdings_list
      summary: Get holdings filtered by portfolio and/or asset type
      parameters:
        - $ref: '#/components/parameters/portfolio_id_query'
        - name: asset_type
          in: query
          schema:
            type: string
            enum: [EQUITY, MUTUAL_FUND, ETF, BOND, GOLD, INTERNATIONAL, OTHER]
      responses:
        "200":
          description: Holdings list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Holding'

    post:
      tags: [portfolio]
      operationId: holdings_add
      summary: Add a new holding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, asset_type, quantity, buy_price, current_price]
              properties:
                name:
                  type: string
                ticker:
                  type: string
                  nullable: true
                asset_type:
                  type: string
                quantity:
                  type: number
                buy_price:
                  type: number
                current_price:
                  type: number
                sector:
                  type: string
                  nullable: true
                buy_date:
                  type: string
                  format: date
                  nullable: true
                portfolio_id:
                  type: string
                  nullable: true
      responses:
        "200":
          description: Created holding
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Holding'
...
  /api/plans/{plan_id}/actions/{action_id}:
    patch:
      tags: [plans]
      operationId: plans_update_action
      summary: Update action status (pending/done/skipped)
      parameters:
        - name: plan_id
          in: path
          required: true
          schema:
            type: string
        - name: action_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [pending, done, skipped]
                completion_note:
                  type: string
                  nullable: true
      responses:
        "200":
          description: Updated plan
...
  /api/portfolio/cas-snapshots:
    get:
      tags: [cas_snapshots]
      operationId: cas_snapshots_list
      summary: List CAS snapshot dates
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 60
        - name: profile_id
          in: query
          schema:
            type: string
            nullable: true
      responses:
        "200":
          description: CAS snapshot list
  /api/portfolio/cas-snapshot:
    get:
      tags: [cas_snapshots]
      operationId: cas_snapshot_get
      summary: Get full CAS snapshot with holdings
      parameters:
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: profile_id
          in: query
          schema:
            type: string
            nullable: true
        - name: include_holdings
          in: query
          schema:
            type: boolean
            default: true
      responses:
        "200":
          description: CAS snapshot
  /api/portfolio/cas-performance:
    get:
      tags: [cas_snapshots]
      operationId: cas_performance
      summary: Portfolio value and health line-chart series
      parameters:
        - name: from_date
          in: query
          schema:
            type: string
            format: date
        - name: to_date
          in: query
          schema:
            type: string
            format: date
        - name: profile_id
          in: query
          schema:
            type: string
            nullable: true
      responses:
        "200":
          description: Performance series
  /api/portfolio/cas-sip-summary:
    get:
      tags: [cas_snapshots]
      operationId: cas_sip_summary
      summary: Monthly SIP bar-chart aggregates
      parameters:
        - name: from_date
          in: query
          schema:
            type: string
            format: date
        - name: to_date
          in: query
          schema:
            type: string
            format: date
        - name: profile_id
          in: query
          schema:
            type: string
            nullable: true
      responses:
        "200":
          description: SIP monthly aggregates
  /api/portfolio/cas-sip-breakdown:
    get:
      tags: [cas_snapshots]
      operationId: cas_sip_breakdown
      summary: Per-fund SIP drill-down for a specific month
      parameters:
        - name: month
          in: query
          required: true
          schema:
            type: string
            pattern: "^\\d{4}-\\d{2}$"
            example: "2024-11"
        - name: profile_id
          in: query
          schema:
            type: string
            nullable: true
      responses:
        "200":
          description: Per-fund breakdown
  /api/portfolio/cas-top-transactions:
    get:
      tags: [cas_snapshots]
      operationId: cas_top_transactions
      summary: Top-N transactions by value
      parameters:
        - name: from_date
          in: query
          schema:
            type: string
            format: date
        - name: to_date
          in: query
          schema:
            type: string
            format: date
        - name: n
          in: query
          schema:
            type: integer
            default: 10
        - name: profile_id
          in: query
          schema:
            type: string
            nullable: true
      responses:
        "200":
          description: Top transactions
  /api/portfolio/cas-snapshot/{snapshot_date}/activate:
    post:
      tags: [cas_snapshots]
      operationId: cas_snapshot_activate
      summary: Load an older CAS snapshot into live holdings
      parameters:
        - name: snapshot_date
          in: path
          required: true
          schema:
            type: string
            format: date
        - name: profile_id
          in: query
          schema:
            type: string
            nullable: true
      responses:
        "200":
          description: Activation result

  # ══════════════════════════════════════════════════════════════════════════
  # CAS Transactions
  # ══════════════════════════════════════════════════════════════════════════
  /api/portfolio/transactions:
    get:
      tags: [cas_transactions]
      operationId: transactions_list
      summary: List CAS transactions
      parameters:
        - name: isin
          in: query
          schema:
            type: string
            nullable: true
        - name: folio
          in: query
          schema:
            type: string
            nullable: true
        - name: limit
          in: query
          schema:
            type: integer
            default: 500
      responses:
        "200":
          description: Transactions with totals
...
  /api/copilot/agents/route:
    post:
      tags: [copilot_agents]
      operationId: copilot_route_intent
      summary: Deterministic intent router — classify message to agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message:
                  type: string
      responses:
        "200":
          description: Routed agent ID and spec
...